+++ src/_commons/StorageContext.js
@@ -7,20 +7,29 @@ const StorageContextProvider = props => {
 
     const { storage, error } = useStorage();
 
-    async function setMyStops( value ) {
-        await storage.setItem( 'myStops', JSON.stringify( value ) );
-    }
-    
-    async function getMyStops() {
-        const myStops = await storage.getItem( 'myStops' );
-        const result = myStops ? JSON.parse( myStops ) : [];
-        return result;
-    }
+    const _setItem = key =>
+        async value => {
+            await storage.setItem( key, JSON.stringify( value ) );
+        }
+
+    const _getItem = ( key, defaultValue ) =>
+        async () => {
+            const value = await storage.getItem( key );
+            return ! value ? defaultValue : JSON.parse( value );
+       }
+
+    const setMyStops = _setItem( 'myStops' );
+    const getMyStops = _getItem( 'myStops', [] );
+
+    const setMyRoutes = _setItem( 'myRoutes' );
+    const getMyRoutes = _getItem( 'myRoutes', [] );
 
     // useEffect( () => console.log( 'StorageContext rendering.' ) );
 
     return (
-        <StorageContext.Provider value={ { storage, error, setMyStops, getMyStops } }>
+        <StorageContext.Provider 
+            value={ { storage, error, setMyStops, getMyStops, setMyRoutes, getMyRoutes } }
+        >
             { props.children }
         </StorageContext.Provider>
     )