+++ src/lines/LineGroups.js
@@ -1,4 +1,5 @@
-import React, { useContext } from 'react';
+import React, { useContext, useEffect } from 'react';
+// import { useIsFocused } from '@react-navigation/native';
 import { StyledView, StyledScrollView } from '../_abstract/Styled';
 import { lineGroupStyles as styles } from './styles';
 import { CacheContext } from '../_commons/CacheContext';
@@ -23,6 +24,12 @@ const parseGroups = data => {
 }
 
 const LineGroups = ( { navigation } ) => {
+
+    // const isFocused = useIsFocused();
+
+    // useEffect( () => {
+    //     alert( 'isFocused' )
+    // } , [ isFocused ] )
 
     const { cache } = useContext( CacheContext );
     const { lines } = cache;