+++ src/my/styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+import patterns from '../_commons/stylePatterns';
+
+const styles = StyleSheet.create( {
+
+    main: {
+        ...patterns.main,
+    },
+
+    list: {
+        ...patterns.list,
+    },
+
+} );
+
+export default styles;