+++ src/stops/Stop.js
@@ -10,7 +10,7 @@ const RowDescrText = StyledText( { style: styles.rowDescrText } );
 
 const Stop = props => {
 
-    const { stop, linesNavigation } = props;
+    const { stop, stopNavNavigation } = props;
 
     const onPress = () => {
         if ( stop.routes.error ) { 
@@ -18,7 +18,7 @@ const Stop = props => {
         }
         stop.arrivals = {};  // clear cache due to always request up to date values
 
-        linesNavigation.navigate( 'StopNav', { stop } );  
+        stopNavNavigation.navigate( 'StopNav', { stop } );  
         // navigation.navigate() passes the parameters to the component throught `props.route.params`
     }