+++ src/stops/Stops.js
@@ -11,7 +11,7 @@ const List = StyledScrollView( { style: styles.list } );
 
 const Stops = props => {
 
-    const { data, linesNavigation } = props;
+    const { data, stopNavNavigation } = props;
     const { route } = data;
     const { stops } = route;
 
@@ -42,7 +42,7 @@ const Stops = props => {
                 <Stop 
                     key={ i }
                     stop={ stop }
-                    linesNavigation={ linesNavigation }
+                    stopNavNavigation={ stopNavNavigation }
                 />
             ) ) } 
             </List>